# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  files.Job:
    properties:
      error:
        type: string
      finishedAt:
        type: string
      id:
        type: string
      kind:
        enum:
        - pull
        - backup
        - restore
        type: string
      label:
        type: string
      startedAt:
        type: string
      state:
        description: State is running, done or failed.
        enum:
        - running
        - done
        - failed
        type: string
    required:
    - id
    - kind
    - label
    - startedAt
    - state
    type: object
...
  httpapi.CreateServerRequest:
    properties:
      cpuMillis:
        example: 2000
        type: integer
      diskMiB:
        example: 10240
        type: integer
      displayName:
        example: Survival
        type: string
      egg:
        example: paper
        type: string
      environment:
        additionalProperties:
          type: string
        type: object
      externalTrafficPolicy:
        allOf:
        - $ref: '#/definitions/v1alpha1.TrafficPolicy'
        description: 'ExternalTrafficPolicy of the service (default Local: the server
          sees the players'' IPs).'
        enum:
        - Local
        - Cluster
      image:
        example: ghcr.io/pelican-eggs/yolks:java_21
        type: string
      loadBalancerIP:
        description: LoadBalancerIP optionally requests a fixed IP from the pool.
        example: 192.168.1.70
        type: string
      loadBalancerPool:
        description: 'LoadBalancerPool the address comes from (default: the default
          of the panel settings).'
        example: general-pool
        type: string
      memoryMiB:
        example: 4096
        type: integer
      owner:
        description: 'Owner is the username the server belongs to (default: the calling
          admin).'
        example: alice
        type: string
      ports:
        example:
        - 25565
        items:
          type: integer
        type: array
      skipInstall:
        type: boolean
      startOnCompletion:
        description: StartOnCompletion starts the server once the installation is
          finished.
        example: true
        type: boolean
      startup:
        type: string
      storageClass:
        description: 'StorageClass of the data volume (default: the default of the
          panel settings).'
        example: longhorn
        type: string
    required:
    - diskMiB
    - displayName
    - egg
    - memoryMiB
    - ports
    type: object
...
  httpapi.ErrorResponse:
    properties:
      error:
        example: server not found
        type: string
      fields:
        additionalProperties:
          type: string
        description: Fields holds validation errors per field or variable.
        type: object
    required:
    - error
    type: object
...
  httpapi.GameServerList:
    properties:
      items:
        items:
          $ref: '#/definitions/v1alpha1.GameServer'
        type: array
    required:
    - items
    type: object
...
  httpapi.JobList:
    properties:
      items:
        items:
          $ref: '#/definitions/files.Job'
        type: array
    required:
    - items
    type: object
...
  httpapi.ServerStats:
    properties:
      cpuMillis:
        description: CPUMillis is the current CPU usage (1000 = one core); null when
          unknown.
        example: 350
        type: integer
        x-nullable: true
      diskBytes:
        description: DiskBytes is the space used on the data volume; null when unknown.
        example: 524288000
        type: integer
        x-nullable: true
      memoryBytes:
        description: MemoryBytes is the current memory usage; null when unknown.
        example: 1073741824
        type: integer
        x-nullable: true
      phase:
        allOf:
        - $ref: '#/definitions/v1alpha1.Phase'
        example: Running
      uptimeSeconds:
        description: UptimeSeconds since the game pod was started.
        example: 3600
        type: integer
    required:
    - cpuMillis
    - diskBytes
    - memoryBytes
    - phase
    - uptimeSeconds
    type: object
...
  httpapi.UpdateServerRequest:
    properties:
      cpuMillis:
        type: integer
      crashRestart:
        type: boolean
      diskMiB:
        type: integer
      displayName:
        type: string
      environment:
        additionalProperties:
          type: string
        type: object
      externalTrafficPolicy:
        allOf:
        - $ref: '#/definitions/v1alpha1.TrafficPolicy'
        description: ExternalTrafficPolicy of the service; admins only.
        enum:
        - Local
        - Cluster
      image:
        type: string
      loadBalancerIP:
        type: string
      loadBalancerPool:
        description: |-
          LoadBalancerPool moves the server to another enabled pool (its address changes).
          Users may change it; a fixed IP is cleared unless a new one is given.
        type: string
      memoryMiB:
        type: integer
      ports:
        items:
          type: integer
        type: array
      startup:
        type: string
      stopTimeoutSeconds:
        type: integer
    type: object
...
  v1alpha1.GameServer:
    properties:
      apiVersion:
        description: |-
          APIVersion defines the versioned schema of this representation of an object.
          Servers should convert recognized schemas to the latest internal value, and
          may reject unrecognized values.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources
          +optional
        type: string
      kind:
        description: |-
          Kind is a string value representing the REST resource this object represents.
          Servers may infer this from the endpoint the client submits requests to.
          Cannot be updated.
          In CamelCase.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds
          +optional
        type: string
      metadata:
        $ref: '#/definitions/v1.ObjectMeta'
      spec:
        $ref: '#/definitions/v1alpha1.GameServerSpec'
      status:
        $ref: '#/definitions/v1alpha1.GameServerStatus'
    required:
    - spec
    type: object
  v1alpha1.GameServerSpec:
    properties:
      crashRestart:
        description: |-
          CrashRestart restarts the server after a crash (not twice within 60 seconds).
          +kubebuilder:default=true
        type: boolean
      displayName:
        description: DisplayName is a human friendly name.
        type: string
      eggRef:
        description: EggRef is the name of the Egg in the same namespace.
        type: string
      environment:
        additionalProperties:
          type: string
        description: |-
          Environment holds values for the egg variables.
          +optional
        type: object
      externalTrafficPolicy:
        allOf:
        - $ref: '#/definitions/v1alpha1.TrafficPolicy'
        description: |-
          ExternalTrafficPolicy of the service; empty means Local.
          +optional
      image:
        description: Image is the runtime image (one of the egg's docker images or
          a custom one).
        type: string
      installRevision:
        description: |-
          InstallRevision triggers a (re)install whenever it differs from status.installedRevision.
          +kubebuilder:default=1
        type: integer
      loadBalancerIP:
        description: |-
          LoadBalancerIP requests a specific IP from the load balancer pool.
          +optional
        type: string
      loadBalancerPool:
        description: |-
          LoadBalancerPool is the Cilium LB IPAM pool the address comes from. The service
          gets the labels of the pool's service selector.
          +optional
        type: string
      ports:
        description: |-
          Ports are published as TCP and UDP. The first one is the primary port (SERVER_PORT).
          +kubebuilder:validation:MinItems=1
        items:
          type: integer
        type: array
      resources:
        $ref: '#/definitions/v1alpha1.Resources'
      schedules:
        description: |-
          Schedules run tasks at fixed times.
          +optional
          +listType=map
          +listMapKey=name
          +kubebuilder:validation:MaxItems=20
        items:
          $ref: '#/definitions/v1alpha1.Schedule'
        type: array
      skipInstall:
        description: |-
          SkipInstall skips the egg install script.
          +optional
        type: boolean
      startup:
        description: |-
          Startup overrides the egg's startup command.
          +optional
        type: string
      state:
        allOf:
        - $ref: '#/definitions/v1alpha1.PowerState'
        description: |-
          State is the desired power state.
          +kubebuilder:default=Stopped
      stopTimeoutSeconds:
        description: |-
          StopTimeoutSeconds is how long to wait after the stop command before killing the server.
          +kubebuilder:default=600
        type: integer
      storageClass:
        description: |-
          StorageClass of the data volume; it cannot be changed once the volume exists.
          +optional
          +kubebuilder:validation:XValidation:rule="self == oldSelf",message="storageClass cannot be changed"
        type: string
      suspended:
        description: |-
          Suspended servers are stopped and cannot be started or used by their owner (admins only).
          +optional
        type: boolean
    required:
    - eggRef
    - image
    - ports
    - resources
    type: object
  v1alpha1.GameServerStatus:
    properties:
      address:
        description: Address is the external IP assigned by the load balancer.
        type: string
      diskMeasuredAt:
        type: string
      diskUsedBytes:
        description: |-
          DiskUsedBytes is the last measured size of the server files; the files pod that
          measures it only runs on demand.
        type: integer
      installExitCode:
        description: InstallExitCode is the exit code of the last install script run.
        type: integer
      installedRevision:
        description: InstalledRevision is the last completed install revision.
        type: integer
      lastCrashAt:
        description: LastCrashAt is the time of the last detected crash.
        type: string
      lastExitCode:
        description: LastExitCode is the exit code of the last terminated game process.
        type: integer
      message:
        type: string
      observedGeneration:
        type: integer
      phase:
        $ref: '#/definitions/v1alpha1.Phase'
      podUID:
        description: PodUID identifies the current game pod.
        type: string
      restartRequired:
        description: |-
          RestartRequired is set while the running game pod was created with other runtime
          settings (image, startup, variables, memory, CPU, ports) than the spec has now.
        type: boolean
      schedules:
        description: |-
          Schedules holds the last run of every schedule.
          +optional
          +listType=map
          +listMapKey=name
        items:
          $ref: '#/definitions/v1alpha1.ScheduleStatus'
        type: array
      startedAt:
        description: StartedAt is when the current game pod was created.
        type: string
      stopRequestedAt:
        description: StopRequestedAt is set when a graceful stop was sent to the current
          pod.
        type: string
      stopTasksStartedAt:
        description: |-
          StopTasksStartedAt is set while the "stopping" schedules of the current pod run; the
          stop is sent when they are done.
          +optional
        type: string
    type: object
...
  v1alpha1.Phase:
    enum:
    - Pending
    - Installing
    - InstallFailed
    - Offline
    - Starting
    - Running
    - Stopping
    type: string
    x-enum-varnames:
    - PhasePending
    - PhaseInstalling
    - PhaseInstallFailed
    - PhaseOffline
    - PhaseStarting
    - PhaseRunning
    - PhaseStopping
  v1alpha1.PowerState:
    enum:
    - Running
    - Stopped
    type: string
    x-enum-varnames:
    - PowerRunning
    - PowerStopped
  v1alpha1.Resources:
    properties:
      cpuMillis:
        description: |-
          CPUMillis limits CPU usage (1000 = one core); 0 means unlimited.
          +optional
        type: integer
      diskMiB:
        description: |-
          DiskMiB is the size of the persistent volume.
          +kubebuilder:validation:Minimum=256
        type: integer
      memoryMiB:
        description: |-
          MemoryMiB is the memory available to the server (SERVER_MEMORY).
          +kubebuilder:validation:Minimum=64
        type: integer
    required:
    - diskMiB
    - memoryMiB
    type: object
  v1alpha1.Schedule:
    properties:
      cron:
        description: |-
          Cron is a five-field cron expression ("0 4 * * *") or a descriptor like "@daily",
          evaluated in the panel's time zone. Either Cron or Event is set.
          +optional
        type: string
      enabled:
        description: Enabled schedules run automatically; disabled ones can still
          be run by hand.
        type: boolean
      event:
        description: |-
          Event runs the schedule when the server was marked as running ("started") or before it
          is stopped or restarted ("stopping"; the stop waits until the tasks are done).
          +optional
          +kubebuilder:validation:Enum=started;stopping
        enum:
        - started
        - stopping
        type: string
      name:
        description: |-
          Name identifies the schedule within the server.
          +kubebuilder:validation:MinLength=1
          +kubebuilder:validation:MaxLength=50
        type: string
      onlyWhenOnline:
        description: |-
          OnlyWhenOnline skips the run when the server is not running.
          +optional
        type: boolean
      tasks:
        description: |-
          +kubebuilder:validation:MinItems=1
          +kubebuilder:validation:MaxItems=10
        items:
          $ref: '#/definitions/v1alpha1.ScheduleTask'
        type: array
    required:
    - enabled
    - name
    - tasks
    type: object
  v1alpha1.ScheduleStatus:
    properties:
      lastResult:
        description: 'LastResult is "ok", "skipped: …" or "failed: …".'
        type: string
      lastRunAt:
        type: string
      name:
        type: string
    required:
    - name
    type: object
...
  v1alpha1.TrafficPolicy:
    enum:
    - Local
    - Cluster
    type: string
    x-enum-varnames:
    - TrafficLocal
    - TrafficCluster
...
info:
  contact: {}
  description: Game server panel for Kubernetes that runs Pterodactyl and Pelican
    eggs.
  title: Kubedactyl API
  version: dev
paths:
...
  /servers:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.GameServerList'
      security:
      - BearerAuth: []
      summary: List game servers
      tags:
      - Servers
    post:
      consumes:
      - application/json
      description: Creates the server and runs the egg install script. Variables are
        validated against the egg rules.
      parameters:
      - description: Server
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/httpapi.CreateServerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v1alpha1.GameServer'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a game server
      tags:
      - Servers
  /servers/{server}:
    delete:
      description: Deletes the server including its data volume.
      parameters:
      - description: Server name
        in: path
        name: server
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete a game server
      tags:
      - Servers
    get:
      parameters:
      - description: Server name
        in: path
        name: server
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1alpha1.GameServer'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a game server
      tags:
      - Servers
    patch:
      consumes:
      - application/json
      description: Runtime settings apply on the next start. Disk can only grow.
      parameters:
      - description: Server name
        in: path
        name: server
        required: true
        type: string
      - description: Changes
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/httpapi.UpdateServerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1alpha1.GameServer'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a game server
      tags:
      - Servers
...
  /servers/{server}/jobs:
    get:
      parameters:
      - description: Server name
        in: path
        name: server
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.JobList'
      security:
      - BearerAuth: []
      summary: Background jobs of a server
      tags:
      - Backups
...
  /servers/{server}/stats:
    get:
      description: |-
        CPU and memory from metrics-server (cached for 10 seconds), disk usage of the server files (cached
        for 30 seconds; the last measured value while neither the file container nor the game runs).
      parameters:
      - description: Server name
        in: path
        name: server
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.ServerStats'
      security:
      - BearerAuth: []
      summary: Live resource usage
      tags:
      - Servers
...
securityDefinitions:
  BearerAuth:
    description: |-
      "Bearer <token>": a session token from /auth/login or an API token (kdt_…). The web UI
      uses the kd_session cookie.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"